Initial cut at multiple-backend supporting shoes exectuable
This uses a scheme where when a Shoes backend is chosen, we write to
bin/.shoes-backend with the specific command to be run by the base shoes
script to start things up.

While not done yet, this will allow us to move the actual executable
that we run into the backend gems. So ruby-shoes will go away in favor
of backend-specific exectuables (probably by convention named after
their gem, so shoes-swt for the first one).

shoes-picker is the heart of selecting a backend from the available
gems, and then ensuring that the .shoes-backend is written out to. This
is expected to only happen once in a given installation, not on every
app run, which is why we can do it even on JRuby with startup time since
it won't get in the way of most process starts.